fix: :bug: Update SEO and language routing
This commit updates the SEO component in the `seo.svelte` file to use the `PUBLIC_HOSTNAME` environment variable for the canonical URL and image URL.
It also adds a new `urlPathName` variable to handle the case when the URL pathname is `/`.
Additionally, it updates the `link` and `openGraph` tags in the `<svelte:head>` section to use the `hostname` and `urlPathName` variables.
In the `+layout.svelte` file, it adds a new `<link>` tag for the Vietnamese language version using the `hostname` and `route` functions. The `manifest.json` file is updated to change the app name to "CoTAI Education".
The `svelte.config.js` and `vite.config.ts` files are modified to remove the unused `loadEnv` function and update the `routes` configuration in the Vite adapter.